+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 * A component's bounding area will stay up to date when changing the 'pixels per unit' in the editor.
 * Zooming the camera in the editor is a lot smoother now as the zoom amount is based on the current view height and not a flat value.
 * Fixed a crash that occurred when loading assets (mostly fonts) before the AssetManager had an instance.
+* Fixed a crash that occurred when the start up scene identifier is null in a project during loading.
 
 ---